feat: export as json
